fix: remove proxy
